refactor(react-components): SetPlanAddTrigger を CreateSetPlanRow にリネームし JSX を分割する
- リネーム: SetPlanAddTrigger → CreateSetPlanRow。CreateProgramCard /
  CreateDayCard と naming pattern を揃え、SetPlanRow* と Row サフィックスで
  並列に見える。「Trigger」は何のトリガーか不明瞭だったため
- JSX 分割: 親（CreateSetPlanRow）は state + payload 計算ロジックを保持し、
  プレビュー行と ↔ ボタンを internal な PreviewButton / CycleButton に分けて
  presentational に切り出す
- RPE デフォルトを 7 から 8 へ変更。RPE 8 は working set としての最頻値で、
  「そのまま使う、チョット調整したいときだけ編集」の前提として妥当
- 型を set-plan-types から import するように整理し、SetPlanSection も
  共通の SetPlanWithParams を Add/Change のペイロードに統一
