+++ packages/react-components/src/views/program-detail/set-plan-row/index.ts
@@ -1,4 +1,4 @@
-export { SetPlanAddTrigger } from "./set-plan-add-trigger";
+export { CreateSetPlanRow } from "./create-set-plan-row";
 export { SetPlanRowEmpty } from "./set-plan-row-empty";
 export { SetPlanRowRepsXRpe } from "./set-plan-row-reps-x-rpe";
 export { SetPlanRowWeightXReps } from "./set-plan-row-weight-x-reps";
